requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
crispy-bootstrap5==0.7
crispy==0.7.4
cryptography==41.0.2
cycler==0.11.0
defusedxml==0.7.1
django-allauth==0.54.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
Django==4.2
djangorestframework==3.14.0
fabio==2023.6.0
factory-boy==3.3.0
Faker==19.3.0
fonttools==4.41.1
h11==0.14.0
h5py==3.9.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pip==23.2
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQt5==5.15.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
requests==2.31.0
setuptools==67.6.1
silx==1.1.2
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.23.2
wheel==0.40.0
